backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
instagrapi==2.1.3
passlib==1.7.4
pillow==11.1.0
pyasn1==0.4.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic-settings==2.8.1
pydantic_core==2.27.1
pyodbc==5.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
